+++ NEWS.md
@@ -1,3 +1,5 @@
+# patchwork (development version)
+
 # patchwork 1.1.3
 
 * `NULL` can now be used with the different arithmetic operators and will result